% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_export_methods.R
\name{all_export_methods}
\alias{all_export_methods}
\title{Список доступных методов API Export B2B Движение}
\usage{
all_export_methods()
}
\description{
Подробное описание методов на  https://dataexport.docs.apiary.io/. 
Функция выводить список достпных методов в формате вектора character
}
