% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_response_functions.R
\name{api_get_response}
\alias{api_get_response}
\title{Get http response from url}
\usage{
api_get_response(url)
}
\arguments{
\item{url}{URL from \code{\link{api_new_url}}}
}
\description{
Get http response from url
}
