% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_url_functions.R
\name{api_new_url}
\alias{api_new_url}
\title{Create new URL for API request}
\usage{
api_new_url(b2b_site, method, limit = 1000, offset = 0, ...)
}
\arguments{
\item{b2b_site}{B2Bmotion site adress ("avs.express", "b2b.el-com.ru" etc)}

\item{method}{method from \code{\link{api_all_methods}}}

\item{limit}{maximum count of elements in export API response}

\item{offset}{how much element to skip}

\item{filters}{character string to filter for current method}
}
\description{
Returns the character link from parameters
}
