% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_response_functions.R
\name{api_tidy_response}
\alias{api_tidy_response}
\title{Tidy response}
\usage{
api_tidy_response(response)
}
\arguments{
\item{response}{http-response from \code{\link{api_get_response}}}
}
\description{
Create tidy tibble from http-response
}
