% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_url_functions.R
\name{api_url_add_filter}
\alias{api_url_add_filter}
\title{Управление фильтрами для запросов к Export API}
\usage{
api_url_add_filter(url, ...)
}
\description{
`api_create_url_filter` создаёт строку (потом заменим на объект S3) для
последующего присоединения к URL для запроса по API:

- для начала мы делаем только в формате character

- TODO: потом нужно оформить объект S3 `api_url_filter`
}
\details{
Стоит обратить внимание, что это просто character
}
