% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_url_functions.R
\name{api_url_set_site}
\alias{api_url_set_site}
\title{Set site to url}
\usage{
api_url_set_site(url, site)
}
\arguments{
\item{url}{URT to change site}

\item{site}{new site}
}
\description{
Set B2Bmotion site adress to url
}
