% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_url.R
\name{as.character.export_filters}
\alias{as.character.export_filters}
\title{Функция для объекта со списком фильтров API Export в строку}
\usage{
\method{as.character}{export_filters}(filters)
}
\arguments{
\item{filters}{объект S3 `export_filters` со списком фильтров}
}
\description{
Функция необходима для создания ссылки API Export
}
