% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_url.R
\name{as.character.export_url}
\alias{as.character.export_url}
\title{Создание текстовой строки из объекта export_url}
\usage{
\method{as.character}{export_url}(url)
}
\arguments{
\item{url}{объект типа `export_url`}
}
\description{
Создание текстовой строки из объекта export_url
}
