% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2b_export.R
\name{b2b_export2}
\alias{b2b_export2}
\title{B2B export functions}
\usage{
b2b_export2(b2b_site, method, limit = 100, ...)
}
\arguments{
\item{b2b_site}{B2Bmotion site adress ("avs.express", "b2b.el-com.ru" etc)}

\item{method}{API method from \code{\link{api_all_methods}}}

\item{limit}{maximum elements in one page}

\item{...}{filters for \code{\link{api_new_url}}}
}
\description{
B2B export functions
}
