% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2b_key_functions.R
\name{b2b_set_key}
\alias{b2b_set_key}
\title{Set export API key}
\usage{
b2b_set_key(b2b_site, b2b_key)
}
\arguments{
\item{b2b_site}{B2Bmotion site adress}

\item{b2b_key}{secret export-API key}
}
\description{
Set export API key
}
