% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2b_key_functions.R
\name{b2b_set_user_api_key}
\alias{b2b_set_user_api_key}
\title{Set user API key}
\usage{
b2b_set_user_api_key(b2b_site, b2b_user_key)
}
\arguments{
\item{b2b_site}{B2Bmotion site adress}

\item{b2b_user_key}{secret user-API key}
}
\description{
Set user API key
}
