% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_url.R
\name{export_filters}
\alias{export_filters}
\title{Созданием фильтра для URL API Export}
\usage{
export_filters(...)
}
\arguments{
\item{...}{список параметров фильтра `id = c(13, 34), createdTo = "2019-01-01"`

нужен пример}
}
\description{
Функция для создания объекта `export_filters`. Этот объект
}
