% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_url.R
\name{export_methods}
\alias{export_methods}
\title{Список доступных методов API Export B2B Движение}
\usage{
export_methods()
}
\description{
Подробное описание методов на  https://dataexport.docs.apiary.io/. 
Функция выводить список достпных методов в формате вектора character
}
