% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_url.R
\name{export_url}
\alias{export_url}
\title{Создание ссылки для доступа API Export B2B Движение}
\usage{
export_url(b2b_site, method, limit = 100, offset = 0, ...)
}
\arguments{
\item{b2b_site}{адрес сайта без "http://" ("b2b.sanergy.ru", "avs.express" и т.д.)}

\item{method}{метод Export API, список всех методов можно получить 
через \code{\link{all_export_methods}}}

\item{limit}{количество запрашиваемых объектов на одной странице (не может 
быть больше 1000), рекомендуется не более 500}

\item{...}{набор фильтров}
}
\description{
Подробное описание Export API по ссылке: https://dataexport.docs.apiary.io/
}
