% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2b_key_functions.R
\name{set_api_export_key}
\alias{set_api_export_key}
\title{Установка ключа для API Export}
\usage{
set_api_export_key(b2b_site, b2b_key)
}
\arguments{
\item{b2b_site}{B2Bmotion site adress}

\item{b2b_key}{secret export-API key}
}
\description{
Установка ключа для API Export
}
